requirements.txt
tweepy>=4.15.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0